+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/fiori-annotation-api
 
+## 0.2.12
+
+### Patch Changes
+
+-   Updated dependencies [42f13eb]
+    -   @sap-ux/project-access@1.28.3
+    -   @sap-ux/cds-odata-annotation-converter@0.3.8
+
+## 0.2.11
+
+### Patch Changes
+
+-   Updated dependencies [8af1ba2]
+    -   @sap-ux/odata-vocabularies@0.4.4
+    -   @sap-ux/cds-annotation-parser@0.2.4
+    -   @sap-ux/cds-odata-annotation-converter@0.3.8
+    -   @sap-ux/xml-odata-annotation-converter@0.3.2
+
 ## 0.2.10
 
 ### Patch Changes